+++ Makefile
@@ -4,11 +4,10 @@ tests:
 	poetry run pytest
 
 tests-basic:
-	poetry run pytest -n auto \
-	tests/test_map.py \
-		tests/test_map_parallel.py \
-		tests/test_filter.py \
-		tests/test_api.py
+	poetry run pytest tests/basic/test_basic_map.py
+	poetry run pytest tests/basic/test_basic_reduce_resolve.py
+	poetry run pytest tests/basic/test_basic_parallel_map.py
+	poetry run pytest tests/basic/test_basic_filter_split_gather.py
 
 lint:
 	poetry run ruff check docetl/* --fix
