+++ docetl/optimizers/utils.py
@@ -1,23 +1,9 @@
-import re
 from typing import Any, Dict, List
 
 from litellm import completion, completion_cost
-from jinja2 import Environment, meta
+
 from docetl.operations.utils import truncate_messages
 from docetl.utils import completion_cost
-
-
-def render_jinja_template(template_string: str, data: Dict[str, Any]) -> str:
-    """
-    Render a Jinja2 template with the given data, ensuring protection against template injection vulnerabilities.
-    If the data is empty, return an empty string.
-    """
-    if not data:
-        return ""
-
-    env = Environment(autoescape=True)
-    template = env.from_string(template_string)
-    return template.render(input=data)
 
 
 class LLMClient:
