+++ docetl/schemas.py
@@ -1,5 +1,6 @@
+from typing import Any, Dict, List, Optional, Union
+
 from pydantic import BaseModel, field_validator
-from typing import Dict, Optional, List, Union, Any
 
 
 class ToolFunction(BaseModel):
