+++ pyproject.toml
@@ -18,15 +18,12 @@ tqdm = "^4.66.4"
 rich = "^13.7.1"
 tenacity = "^9.0.0"
 frozendict = "^2.4.4"
-scikit-learn = "^1.5.1"
 diskcache = "^5.6.3"
-textual = "^0.79.1"
 typer = "^0.12.5"
-transformers = "^4.44.2"
 pydantic = "^2.9.2"
-pytest-sugar = "^1.0.0"
-pytest-xdist = "^3.6.1"
 asteval = "^1.0.4"
+scikit-learn = "^1.5.2"
+numpy = "^1.26.4"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8.3.2"
@@ -41,6 +38,8 @@ linkchecker = "^10.5.0"
 pytkdocs = "^0.16.2"
 mkdocstrings-python = "^1.11.1"
 mkdocs-glightbox = "^0.4.0"
+pytest-sugar = "^1.0.0"
+pytest-xdist = "^3.6.1"
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
