+++ tests/test_reduce_scale.py
